+++ lib/main.dart
@@ -8,9 +8,9 @@
 
 import 'package:logging/logging.dart';
 import 'package:ydits_ssc/features/app_runner/app_runner.dart';
-import 'package:ydits_ssc/features/sub_apps/runner/sub_app_runner.dart';
+import 'package:ydits_ssc/features/app_runner/sub_app_runner.dart';
 import 'package:ydits_ssc/features/main_app/configure/main_app_window_config.dart';
-import 'package:ydits_ssc/features/main_app/runner/main_app_runner.dart';
+import 'package:ydits_ssc/features/app_runner/main_app_runner.dart';
 
 late final Logger logger;
